+++ Assets/Scripts/Fundamental/Player/PlayerWeaponCTRL.cs
@@ -9,7 +9,7 @@ public class PlayerWeaponCTRL : MonoBehaviour
     [SerializeField] private RaycastCTRL _raycastCtrl;
     private GameObject activeWeaponShown;
     private WeaponScript activeWeaponScript;
-    private float bulletDamage;
+    private int bulletDamage;
     private float timer = 1;
     
     public void InitiaterWeaponCtrl()
